requirements.txt
# Computer Vision
opencv-python
mediapipe

# Web & UI
streamlit
flask

# Backend & Database
firebase-admin
pymongo[srv]

# AI & LLM
langchain
google-generativeai
python-dotenv

pytz